[POC] Move compressor impl to pkg/trace
